pyproject.toml
[project]
name = "papermate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "fastmcp>=0.4.1",
    "mcp[cli]>=1.6.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
]
